♻️ refactor: contants 폴더명 변경
